Moved listener classes to folder ui.auxiliary
